feat: monitor layout
